Cleaned up api/create-rest-api
api/create-rest-api now has 3 different arities.

(api/create-rest-api) - Looks in the environment for :username,
:password, :rally-host and :api-key. See
https://github.com/weavejester/environ for more information.

(api/create-rest-api credentials) - credentials is a map that contains
either a :username and :password or :api-key. :rally-host is looked up
in the environment or defaults to https://rally1.rallydev.com

(api/create-rest-api credentials rally-host) - This is the most specific
of the arities. All information needed to talk to Rally is required.